market_data_connector/requirements.txt
numpy
grpcio
protobuf
requests
sqlalchemy
pymysql
cryptography
schedule